refactor: decouple registration validation from relay
