+++ mev-relay-rs/src/lib.rs
@@ -1,5 +1,4 @@
 mod relay;
 mod service;
-pub(crate) mod validator_summary_provider;
 
 pub use service::{Config, Service};
