+++ mev-rs/Cargo.toml
@@ -8,13 +8,14 @@ license = "MIT OR Apache-2.0"
 
 [features]
 default = ["serde", "api"]
-api = ["tokio", "axum", "hyper","beacon-api-client", "tracing"]
+api = ["tokio", "axum", "hyper", "beacon-api-client", "tracing"]
 
 [dependencies]
 tokio = { version = "1.0", optional = true }
-axum = { version =  "0.5.4", optional = true }
+axum = { version = "0.5.4", optional = true }
 hyper = { version = "0.14", optional = true }
 tracing = { version = "0.1", optional = true }
+parking_lot = "0.12.1"
 
 async-trait = "0.1.53"
 serde = { version = "1.0", features = ["derive"], optional = true }
