+++ mev-rs/src/lib.rs
@@ -4,11 +4,20 @@ mod network;
 pub mod serde;
 pub mod signing;
 pub mod types;
+pub mod validator_registration;
 
 pub use blinded_block_provider::{BlindedBlockProvider, Error as BlindedBlockProviderError};
 #[cfg(feature = "api")]
 pub use blinded_block_provider::{
     Client as BlindedBlockProviderClient, Server as BlindedBlockProviderServer,
+};
+pub use validator_registration::{
+    validator_registrar::{
+        Error as ValidatorRegistrationError, ValidatorRegistrar, ValidatorRegistrationStatus,
+    },
+    validator_summary_provider::{
+        Error as ValidatorSummaryProviderError, ValidatorSummaryProvider,
+    },
 };
 
 pub use network::*;
