+++ mev-rs/src/validator_registration/mod.rs
@@ -1,0 +1,2 @@
+pub mod validator_registrar;
+pub mod validator_summary_provider;
